Vento syntax highlighting for plain text and HTML
This sets a basis for decent syntax highlighting for Vento, both for
plain text (effectively only highlighting the Vento syntax) and HTML.

Two hidden sublanguages are defined, for JS expressions in Vento and for
JS code in Vento in general. The former is needed so that e.g. `foo |>
bar` doesn't interpret the `|` as a bitwise operator and `>` as a
comparison operator. That would probably highlight the same, but it's
not pretty, especially because in some cases we explicitly have to mark
the pipe as a pipe (for example in `{{ echo |> md }}`. Also, expressions
must be aware that they can end with `-}}`, which is necessary for the
`-` not to be seen as subtraction. The latter, general JS code in Vento,
is needed for the same reason; when doing `{{> let foo = 1 -}}`, the `-`
mustn't be highlighted as an operator but rather as part of the closing
tag.

The HTML definition is rather simplistic right now, with no support for
Vento-in-JS or Vento-in-CSS. This is probably fine in most cases, but it
should be supported in the future anyway.

Also, it should be noted that I have very little experience with Sublime
plugins, so these languages have just been tested as custom syntaxes in
my personal environment. It should probably be made into a plugin hosted
on https://packagecontrol.io/ so people can install it through Sublime
itself.